make requirements create and show commands
