+++ cmd/publisher/main.go
@@ -22,12 +22,12 @@ type cliSpec struct {
 	cli_types.CommonArgs
 	commands.AccountCommands `group:"Accounts"`
 
-	Init         commands.InitCommand         `kong:"cmd" help:"Create a configuration file based on the contents of the project directory."`
-	Deploy       commands.DeployCmd           `kong:"cmd" help:"Create a new deployment."`
-	Redeploy     commands.RedeployCmd         `kong:"cmd" help:"Update an existing deployment."`
-	UI           commands.UICmd               `kong:"cmd" help:"Serve the publisher UI."`
-	Requirements commands.RequirementsCommand `kong:"cmd" help:"Create a Python requirements.txt file."`
-	Version      commands.VersionCmd          `kong:"cmd" help:"Show the client software version and exit."`
+	Init         commands.InitCommand          `kong:"cmd" help:"Create a configuration file based on the contents of the project directory."`
+	Deploy       commands.DeployCmd            `kong:"cmd" help:"Create a new deployment."`
+	Redeploy     commands.RedeployCmd          `kong:"cmd" help:"Update an existing deployment."`
+	UI           commands.UICmd                `kong:"cmd" help:"Serve the publisher UI."`
+	Requirements commands.RequirementsCommands `kong:"cmd" help:"Create a Python requirements.txt file."`
+	Version      commands.VersionCmd           `kong:"cmd" help:"Show the client software version and exit."`
 }
 
 func logVersion(log logging.Logger) {
